Expose the Overscan option to the video menu and fix a potential bug on the game consoles
So it turns out that the overscan code works well on the PC since it is fairly generic, we expose it because there may be a time that someone may want to run the client on a TV using a non-console.
Replace the magic number used in M_OptInit with VID_FSLINE and set it to the correct value, this fixes a bug that if a fullscreen only or window only display capability was present, it would set the wrong menu item to be nochoice (unchangeable by the user), only do this on non-consoles.